+++ bob-common/src/configs/node.rs
@@ -492,6 +492,8 @@ pub struct Node {
     #[serde(default = "Node::default_http_api_address")]
     http_api_address: IpAddr,
     bind_to_ip_address: Option<SocketAddr>,
+    #[serde(default = "NodeConfig::default_filter_group_size")]
+    filter_group_size: usize,
 }
 
 impl NodeConfig {
@@ -654,6 +656,11 @@ impl NodeConfig {
         self.disk_access_par_degree
     }
 
+    #[inline]
+    pub fn filter_group_size(&self) -> usize {
+        self.filter_group_size
+    }
+
     fn check_unset(&self) -> Result<(), String> {
         if self.backend_type == PLACEHOLDER
             || self.check_interval == PLACEHOLDER
@@ -696,6 +703,10 @@ impl NodeConfig {
 
     fn default_http_api_address() -> IpAddr {
         IpAddr::V4(Ipv4Addr::UNSPECIFIED)
+    }
+
+    pub fn default_filter_group_size() -> usize {
+        8
     }
 }
 
@@ -775,6 +786,7 @@ pub mod tests {
             http_api_address: NodeConfig::default_http_api_address(),
             bind_to_ip_address: None,
             bloom_filter_memory_limit: None,
+            filter_group_size: 8,
         }
     }
 }