+++ CHANGELOG.md
@@ -4,7 +4,6 @@ Bob versions changelog
 
 ## [Unreleased]
 #### Added
-- Add bloom_filter_memory_limit to example config (#378)
 - Hierarchical filters (#333)
 
 
@@ -16,6 +15,15 @@ Bob versions changelog
 
 #### Updated
 
+
+## [2.0.0-alpha.10] - 2021-12-02
+#### Added
+- Add bloom_filter_memory_limit to example config (#378)
+- Add traits to support pearl #123
+
+
+#### Updated
+- Upgrade pearl to v0.8.1
 
 
 ## [2.0.0-alpha.9] - 2021-11-19